Extract WallabagDbOpenHelper
